+++ packages/plugin-progress/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @spoosh/plugin-progress
 
+## 0.1.1
+
+- Renamed `PluginContext.requestOptions` to `PluginContext.request` for cleaner API
+
 ## 0.1.0
 
 - Initial release